requirements.txt
crewai[tools]>=0.152.0
ib_insync>=0.9.86
yfinance>=0.2.31
pandas>=2.0
numpy>=1.24
feedparser>=6.0
requests>=2.31
python-telegram-bot>=20.0
psutil>=5.9
schedule>=1.2
python-dotenv>=1.0
pydantic>=2.0
pytest>=7.0
